feat: implement create/update endpoints and adjust service validations

+++ src/main/resources/db/migration/V1__Create_Table_User.sql
@@ -1,5 +1,5 @@
 Create table `users`(
-    `id` int primary key auto_increment,
+    `id` BIGINT primary key auto_increment,
     `first_name` varchar(100) not null,
     `last_name` varchar(120) not null,
     `email` varchar(200) unique not null,
